### 2023-06-09 #5529 @editorialbot
[PRE REVIEW]: Bidsme: expandable bidsifier of brain imagery datasets
**Submitting author:** @nbeliy (<a href="http://orcid.org/0009-0002-0830-3279">Nikita Beliy</a>)
**Repository:** https://github.com/CyclotronResearchCentre/bidsme
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** 1.4.0.post4
