### 2023-06-13 #5529 @nbeliy
Hi @emdupre, thanks for accepting the editing.

From the list [@suyashdb](https://github.com/suyashdb), [@canimus](https://github.com/canimus) [@lparsons](https://github.com/lparsons), [@peerherholz](https://github.com/peerherholz)looking matching profiles. I essentially looked into tags `data formats`, `workflow`, `bioinformatics` and `neuroscience`.

Also [@remi-gau](https://github.com/Remi-Gau) could accept, as he working extensively on BIDS. We collaborated on some projects (just in case if it counts as conflict of interest). He can also suggest some reviewers

A small question, is it possible to add to the pdf aside the link to repository  links to tutorial? And if it is possible, how to do it?
