### 2023-06-13 #5529 @emdupre
Thanks for these suggestions, @nbeliy ! I'll consider them alongside my own research and send out invitations shortly.

To confirm, do you mean the PDF sidebar where e.g. the editor and DOI are listed ? If so, we don't support changing that. That's standardized JOSS formatting and accompanies all papers. If you meant something else, though, please let me know !
