### 2023-06-19 #5529 @emdupre
👋 Hi @TheChymera, @suyashdb,

Would you be willing to review _Bidsme: expandable bidsifier of brain imagery datasets_ for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org/)?

You can learn more about reviewing for JOSS -- including our conflict of interest policy -- [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The review takes place on GitHub and focuses on [the software](https://github.com/CyclotronResearchCentre/bidsme) and this short paper:
https://github.com/openjournals/joss-papers/blob/joss.05529/joss.05529/10.21105.joss.05529.pdf

If you're available to review this work, please let me know and I'll add you as a reviewer. Once I have sufficient reviewers I'll open a dedicated review issue.

Thank you for considering !
