### 2023-06-30 #5575 @emdupre
👋 Hi @PeerHerholz and @TheChymera, and happy Friday !

I just wanted to check whether you're having any issues in getting started with this review.  Please let me know if so ; otherwise, thank you again for agreeing to review _Bidsme_ !
