### 2023-07-08 CyclotronResearchCentre/bidsme#4 @PeerHerholz
JOSS review - installation & general structure
Ahoi hoi @nbeliy & co-authors,

this is in reference to the review I'm currently conducting for `JOSS` [here](https://github.com/openjournals/joss-reviews/issues/5575) .

First of all, thank you very much for your dedicated work on this tool and your respective submission to `JOSS`.
Based on the [review process and included checkboxes](https://github.com/openjournals/joss-reviews/issues/5575#issuecomment-1627300948), I would like to suggest a few things.

When I tried to install `bidsme` following the provided [installation guide](https://gitlab.uliege.be/CyclotronResearchCentre/Public/bidstools/bidsme/bidsme_tutorial/-/tree/master/Installation), I ran into an error saying:

`ModuleNotFoundError: No module named 'ruamel'`

and thus, I could not evaluate the other tutorials and performance. It might also be nice to add your package to [pypi](https://pypi.org/).

Furthermore, it might be beneficial to have everything, ie the software, documentation and examples, in the same `github` repository and split between `github` and `gitlab` as this potentially complicates things for interested users (ie finding information, etc.).
Related to that: the [README file](https://github.com/CyclotronResearchCentre/bidsme/blob/dev/README.md) could be enhanced by adding more information, e.g. regarding the installation, etc. .

I'm very sorry if I missed something obvious, misunderstood something or assumed something wrong. Please let me know if that is the case, as well as if you have any questions. Thanks again for creating this important software tool.

Cheers, Peer
