### 2023-07-08 CyclotronResearchCentre/bidsme#7 @PeerHerholz
JOSS review - automated tests
Ahoi hoi @nbeliy & co-authors,

this is in reference to the review I'm currently conducting for `JOSS` [here](https://github.com/openjournals/joss-reviews/issues/5575) .

First of all, thank you very much for your dedicated work on this tool and your respective submission to `JOSS`.
Based on the [review process and included checkboxes](https://github.com/openjournals/joss-reviews/issues/5575#issuecomment-1627300948), I would like to suggest a few things.

It would be great if you could add a few automated tests/CI to the repository, e.g. a little [github workflow](https://docs.github.com/en/actions/using-workflows) that tests the installation and follows/implements the example you provide.

I'm very sorry if I missed something obvious, misunderstood something or assumed something wrong. Please let me know if that is the case, as well as if you have any questions. Thanks again for creating this important software tool.

Cheers, Peer
