### 2023-07-10 CyclotronResearchCentre/bidsme#4 @nbeliy
Hi @PeerHerholz ,

- Can you tell how you tried to install `bidsme` exactly? I crosschecked, `ruamel.yaml` is both in `setup.py` and `requirements.txt`, so it should be installed automatically by pip. Maybe I missed a parameter sonwhere...
- I used to have README to contain all information, but it was too bloated and difficult to find anything in it. I'll check if I can add more revelant links to it.
- `github` and `gitlab` -- will fix it
