### 2023-07-10 CyclotronResearchCentre/bidsme#4 @PeerHerholz
Hi @nbeliy,

thx for the reply and information/explanations!

> Can you tell how you tried to install bidsme exactly? I crosschecked, ruamel.yaml is both in setup.py and requirements.txt, so it should be installed automatically by pip. Maybe I missed a parameter sonwhere...

I followed the provided [installation guide](https://gitlab.uliege.be/CyclotronResearchCentre/Public/bidstools/bidsme/bidsme_tutorial/-/tree/master/Installation).

> I used to have README to contain all information, but it was too bloated and difficult to find anything in it. I'll check if I can add more revelant links to it.

Totally understandable, thanks!

> github and gitlab -- will fix it

Awesome, thx!

Cheers, Peer
