### 2023-07-10 CyclotronResearchCentre/bidsme#4 @nbeliy
Hi Peer,

> I followed the provided [installation guide](https://gitlab.uliege.be/CyclotronResearchCentre/Public/bidstools/bidsme/bidsme_tutorial/-/tree/master/Installation).

That worries me -- I just tried in new environment, and ruamel was installed:
 > Successfully built bidsme
Installing collected packages: pytz, tzdata, ruamel.yaml.clib, numpy, humanfriendly, ruamel.yaml, pandas, coloredlogs, bidsme
Successfully installed bidsme-1.4.0.post4 coloredlogs-15.0.1 humanfriendly-10.0 numpy-1.24.4 pandas-2.0.3 pytz-2023.3 ruamel.yaml-0.17.32 ruamel.yaml.clib-0.2.7 tzdata-2023.3

Can you send me the output of the `pip install` command in tutorial? Maybe a jupyther issue...
