### 2023-07-10 CyclotronResearchCentre/bidsme#4 @PeerHerholz
Hi @nbeliy,

the `pip install` works fine. I get the error when I run `import bidsme`.
Sorry, I should've been more precise.
