### 2023-07-11 CyclotronResearchCentre/bidsme#4 @nbeliy
> the pip install works fine. I get the error when I run import bidsme.

Yes, I understood. The `ruamel` module is supposed to  be installed by pip. Can you confirm that you run bidsme in the same environment where you did `pip install`?

Also, it may be version problem -- if in the later ruamel they renamed the module. Can you give me the python version and ruamel version (you can see it using `pip show  ruamel.yaml`)
