### 2023-07-12 c081acc
Updated entities for pet, removed warning for undecoded strings
