### 2023-07-12 4a04649
Non-bids modalities are accepted with a warning
