### 2023-07-12 a09e5c8
Removed continue from non-Bids modality section
