+++ bidsme/version.txt
@@ -1,1 +1,1 @@
-1.4.0.post4
+1.4.1
