### 2023-07-12 CyclotronResearchCentre/bidsme#4 @nbeliy
Hi, I added the installation section into README and fixed the links, they points to github now.

Will check the community guides later.

Is the situation with ruamel got clarified, or the issue persists?
