### 2023-07-12 CyclotronResearchCentre/bidsme#4 @PeerHerholz
Hi @nbeliy,

sorry for the late reply.

>  Hi, I added the installation section into README and fixed the links, they points to github now.

> Will check the community guides later.

Coolio!

> Yes, I understood. The ruamel module is supposed to be installed by pip. Can you confirm that you run bidsme in the same environment where you did pip install?

Yes, it's the same environment.

> Also, it may be version problem -- if in the later ruamel they renamed the module. Can you give me the python version and ruamel version (you can see it using pip show  ruamel.yaml)

I tested it on `python 3.6` and the `ruamel` version I get is `0.15.71`.
