### 2023-07-12 CyclotronResearchCentre/bidsme#4 @nbeliy
> I tested it on python 3.6 and the ruamel version I get is 0.15.71

Weird, I tried into a new environment, and has no problem with yaml. Out of desperation, can you try following commands, and send me output?
```bash
python3.6 -m venv bidsme-test
source bidsme-test/bin/activate
python3 -m pip install git+https://github.com/CyclotronResearchCentre/bidsme.git
bidsme
```
