### 2023-07-13 cec2cae
Extra fields in tables json files
