### 2023-07-14 CyclotronResearchCentre/bidsme#4 @PeerHerholz
Hi @nbeliy,

sorry for the late reply.
Creating a new environment resolves the issue and the `CLI` and `python` usage options work. I'm wondering what happened in the environment I tried before. Maybe it's worthwhile adding a small note concerning environments (e.g. "It is recommended to create and utilize a dedicated python environment.")?
