### 2023-07-27 CyclotronResearchCentre/bidsme#9 @TheChymera
JOSS Review — Versioning scheme
As part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/5575

![decohost_2023-07-27T00:44:36](https://github.com/CyclotronResearchCentre/bidsme/assets/950524/ae96ac3e-f838-47ca-b3e8-7146d123a6fa)

I would recommend a versioning scheme that is somewhat cleaner. `major.minor.patch` is best. While I have seen the `post` and `pre` suffixes before, they usually reference a date or some more meaningful information that couldn't otherwise be expressed in the major-minor-patch format. `post4` sound like it should have been a patch, i.e. `1.4.4`.

Not least of all it causes problems for semantic parsing of the version number.
