### 2023-07-27 CyclotronResearchCentre/bidsme#10 @TheChymera
JOSS Review — Test suite
As part of the JOSS Review → https://github.com/openjournals/joss-reviews/issues/5575

This is analogous with the issue raised here https://github.com/CyclotronResearchCentre/bidsme/issues/7

I fully appreciate that it can be difficult to formulate exhaustive or even just comprehensive tests for scientific software, but without a test suite it is impossible for me to determine whether the software can do anything at all.

The other reviewer has mentioned github workflows, which are a very useful feature. However if for some reason you don't want to do that, I'd be satisfied with just a few tests I can run locally, covering some of the functionalities mentioned in the article.
