### 2023-07-27 CyclotronResearchCentre/bidsme#11 @TheChymera
JOSS Review — Installation Proof
As part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/5575

Could you confirm that the following files constitute a successful installation of the package?

```
decohost ~ # equery f bidsme
 * Searching for bidsme ...
 * Contents of dev-python/bidsme-1.4.4:
/usr
/usr/bin
/usr/bin/bidsme -> ../lib/python-exec/python-exec2
/usr/bin/bidsme-pdb -> ../lib/python-exec/python-exec2
/usr/lib
/usr/lib/python-exec
/usr/lib/python-exec/python3.11
/usr/lib/python-exec/python3.11/bidsme
/usr/lib/python-exec/python3.11/bidsme-pdb
/usr/lib/python3.11
/usr/lib/python3.11/site-packages
/usr/lib/python3.11/site-packages/bidsme
/usr/lib/python3.11/site-packages/bidsme-1.4.0.post4.dist-info
/usr/lib/python3.11/site-packages/bidsme-1.4.0.post4.dist-info/METADATA
/usr/lib/python3.11/site-packages/bidsme-1.4.0.post4.dist-info/RECORD
/usr/lib/python3.11/site-packages/bidsme-1.4.0.post4.dist-info/WHEEL
/usr/lib/python3.11/site-packages/bidsme-1.4.0.post4.dist-info/entry_points.txt
/usr/lib/python3.11/site-packages/bidsme-1.4.0.post4.dist-info/top_level.txt
/usr/lib/python3.11/site-packages/bidsme/Modules
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/BrainVision.py
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/EDF.py
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/EEG.py
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/_EDF.py
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/_EEG.py
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/_MNE.py
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__init__.py
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/BrainVision.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/BrainVision.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/BrainVision.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/EDF.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/EDF.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/EDF.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/EEG.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/EEG.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/EEG.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/_EDF.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/_EDF.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/_EDF.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/_EEG.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/_EEG.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/_EEG.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/_MNE.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/_MNE.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/_MNE.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/__init__.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/__init__.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/EEG/__pycache__/__init__.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/DICOM.py
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/MRI.py
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/NIFTI.py
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/_DICOM.py
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/_MRI.py
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__init__.py
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/DICOM.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/DICOM.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/DICOM.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/MRI.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/MRI.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/MRI.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/NIFTI.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/NIFTI.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/NIFTI.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/_DICOM.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/_DICOM.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/_DICOM.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/_MRI.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/_MRI.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/_MRI.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/__init__.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/__init__.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/__init__.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/_hmriNIFTI.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/_hmriNIFTI.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/_hmriNIFTI.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/bidsmeNIFTI.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/bidsmeNIFTI.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/bidsmeNIFTI.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/hmriNIFTI.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/hmriNIFTI.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/hmriNIFTI.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/jsonNIFTI.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/jsonNIFTI.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/__pycache__/jsonNIFTI.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/_hmriNIFTI.py
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/bidsmeNIFTI.py
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/hmriNIFTI.py
/usr/lib/python3.11/site-packages/bidsme/Modules/MRI/jsonNIFTI.py
/usr/lib/python3.11/site-packages/bidsme/Modules/PET
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/DICOM.py
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/ECAT.py
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/NIFTI.py
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/PET.py
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/_DICOM.py
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/_ECAT.py
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/_PET.py
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__init__.py
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/DICOM.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/DICOM.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/DICOM.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/ECAT.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/ECAT.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/ECAT.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/NIFTI.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/NIFTI.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/NIFTI.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/PET.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/PET.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/PET.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/_DICOM.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/_DICOM.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/_DICOM.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/_ECAT.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/_ECAT.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/_ECAT.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/_PET.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/_PET.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/_PET.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/__init__.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/__init__.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/__init__.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/bidsmeNIFTI.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/bidsmeNIFTI.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/bidsmeNIFTI.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/jsonNIFTI.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/jsonNIFTI.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/__pycache__/jsonNIFTI.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/bidsmeNIFTI.py
/usr/lib/python3.11/site-packages/bidsme/Modules/PET/jsonNIFTI.py
/usr/lib/python3.11/site-packages/bidsme/Modules/__init__.py
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/__init__.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/__init__.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/__init__.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/_constants.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/_constants.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/_constants.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/_dicom_common.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/_dicom_common.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/_dicom_common.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/_nifti_common.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/_nifti_common.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/_nifti_common.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/abstract.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/abstract.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/abstract.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/base.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/base.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/base.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/common.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/common.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/common.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/exceptions.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/exceptions.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/exceptions.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/selector.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/selector.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/__pycache__/selector.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/_constants.py
/usr/lib/python3.11/site-packages/bidsme/Modules/_dicom_common.py
/usr/lib/python3.11/site-packages/bidsme/Modules/_formats
/usr/lib/python3.11/site-packages/bidsme/Modules/_formats/MNE.py
/usr/lib/python3.11/site-packages/bidsme/Modules/_formats/_MNE.py
/usr/lib/python3.11/site-packages/bidsme/Modules/_formats/__init__.py
/usr/lib/python3.11/site-packages/bidsme/Modules/_formats/__pycache__
/usr/lib/python3.11/site-packages/bidsme/Modules/_formats/__pycache__/MNE.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/_formats/__pycache__/MNE.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/_formats/__pycache__/MNE.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/_formats/__pycache__/_MNE.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/_formats/__pycache__/_MNE.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/_formats/__pycache__/_MNE.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/_formats/__pycache__/__init__.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/_formats/__pycache__/__init__.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/_formats/__pycache__/__init__.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/_formats/__pycache__/dummy.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/_formats/__pycache__/dummy.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/_formats/__pycache__/dummy.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/Modules/_formats/dummy.py
/usr/lib/python3.11/site-packages/bidsme/Modules/_nifti_common.py
/usr/lib/python3.11/site-packages/bidsme/Modules/abstract.py
/usr/lib/python3.11/site-packages/bidsme/Modules/base.py
/usr/lib/python3.11/site-packages/bidsme/Modules/common.py
/usr/lib/python3.11/site-packages/bidsme/Modules/exceptions.py
/usr/lib/python3.11/site-packages/bidsme/Modules/selector.py
/usr/lib/python3.11/site-packages/bidsme/__init__.py
/usr/lib/python3.11/site-packages/bidsme/__pycache__
/usr/lib/python3.11/site-packages/bidsme/__pycache__/__init__.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/__pycache__/__init__.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/__pycache__/__init__.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/__pycache__/bidsify.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/__pycache__/bidsify.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/__pycache__/bidsify.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/__pycache__/main.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/__pycache__/main.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/__pycache__/main.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/__pycache__/mapper.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/__pycache__/mapper.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/__pycache__/mapper.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/__pycache__/prepare.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/__pycache__/prepare.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/__pycache__/prepare.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/__pycache__/process.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/__pycache__/process.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/__pycache__/process.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsMeta
/usr/lib/python3.11/site-packages/bidsme/bidsMeta/BidsMeta.py
/usr/lib/python3.11/site-packages/bidsme/bidsMeta/BidsSession.py
/usr/lib/python3.11/site-packages/bidsme/bidsMeta/BidsTable.py
/usr/lib/python3.11/site-packages/bidsme/bidsMeta/__init__.py
/usr/lib/python3.11/site-packages/bidsme/bidsMeta/__pycache__
/usr/lib/python3.11/site-packages/bidsme/bidsMeta/__pycache__/BidsMeta.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsMeta/__pycache__/BidsMeta.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsMeta/__pycache__/BidsMeta.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsMeta/__pycache__/BidsSession.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsMeta/__pycache__/BidsSession.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsMeta/__pycache__/BidsSession.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsMeta/__pycache__/BidsTable.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsMeta/__pycache__/BidsTable.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsMeta/__pycache__/BidsTable.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsMeta/__pycache__/__init__.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsMeta/__pycache__/__init__.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsMeta/__pycache__/__init__.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsify.py
/usr/lib/python3.11/site-packages/bidsme/bidsmap
/usr/lib/python3.11/site-packages/bidsme/bidsmap/__init__.py
/usr/lib/python3.11/site-packages/bidsme/bidsmap/__pycache__
/usr/lib/python3.11/site-packages/bidsme/bidsmap/__pycache__/__init__.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsmap/__pycache__/__init__.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsmap/__pycache__/__init__.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsmap/__pycache__/_bidsmap.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsmap/__pycache__/_bidsmap.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsmap/__pycache__/_bidsmap.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsmap/__pycache__/_run.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsmap/__pycache__/_run.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsmap/__pycache__/_run.cpython-311.pyc
/usr/lib/python3.11/site-packages/bidsme/bidsmap/_bidsmap.py
/usr/lib/python3.11/site-packages/bidsme/bidsmap/_run.py
/usr/lib/python3.11/site-packages/bidsme/bidsversion.txt
/usr/lib/python3.11/site-packages/bidsme/exceptions
/usr/lib/python3.11/site-packages/bidsme/exceptions/__init__.py
/usr/lib/python3.11/site-packages/bidsme/exceptions/__pycache__
/usr/lib/python3.11/site-packages/bidsme/exceptions/__pycache__/__init__.cpython-311.opt-1.pyc
/usr/lib/python3.11/site-packages/bidsme/exceptions/__pycache__/__init__.cpython-311.opt-2.pyc
/usr/lib/python3.11/site-packages/bidsme/exceptions/__pycache__/__init__.cpython-311.pyc
[…]
