### 2023-07-27 CyclotronResearchCentre/bidsme#12 @TheChymera
JOSS Review — Quality of writing
As part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/5575

## Suggestions:
1. Please consider splitting the source code over more lines, ideally one line per sentence to make things easier to reference.
2. Please consider using the term “directories” instead of “folders”. Directories are file system entities, whereas folders are graphical interface features.

## Minor issues:
1. Python is a proper noun and should be capitalized, at least in formal writing; python is a type of snake.

## Major issues:
1. “imposing only the core of BIDS -- the folder structure, the file naming style and the global metadata”. The term “global metadata” is unclear so you would need to introduce and explain it. To my knowledge, “global metadata” is not a core concept in BIDS. I would guess you mean simply `dataset_description.json` — but then again, you do not show this file, but you show e.g. `participants.tsv`, a different metadata file, in the figure.
