### 2023-08-06 CyclotronResearchCentre/bidsme#11 @nbeliy
Looks correct. Except this part (if these are the links):
````
/usr/bin/bidsme -> ../lib/python-exec/python-exec2
/usr/bin/bidsme-pdb -> ../lib/python-exec/python-exec2
```

`bidsme` and `bidsme-pdb` are supposed to point to `/usr/lib/python-exec/python3.11/bidsme` and `/usr/lib/python-exec/python3.11/bidsme-pdb`

Usually I test installation using `bidsme --help`.
