### 2023-08-06 CyclotronResearchCentre/bidsme#12 @nbeliy
Hi @TheChymera

Thanks for the comments.

> “imposing only the core of BIDS -- the folder structure, the file naming style and the global metadata”

Under global I meant here the column `participant_id` in `participants.tsv` and `*_scans.tsv`. Will it be better to remove of mention of `global metadata` completely?

> Python is a proper noun and should be capitalized

Good point

>  Please consider using the term “directories” instead of “folders”.

Will do

> Please consider splitting the source code over more lines,

Can you provide more details/examples? I used flake8 which warns about long lines and usage of multiple directives on same line -- so normally the source code already splitted over multiple lines.
