### 2023-08-06 CyclotronResearchCentre/bidsme#9 @nbeliy
The `pre`/`post` are the versioning style of pip, marking post-release (minor patches).

But you are right, will use the `1.4.4` looks nicer. Will adapt for the next release.
