### 2023-08-06 CyclotronResearchCentre/bidsme#10 @nbeliy
The code is tested [gitlab-ci](https://github.com/CyclotronResearchCentre/bidsme/blob/dev/.gitlab-ci.yml) but on private gitlab of ULiege, which is the main platform. The CRC GitHub is a sort of front-end for publically aviable projects. But I guess I can replicate the same test on GitHub.
