+++ bidsme/.github/workflows/flake8.yml
@@ -1,0 +1,25 @@
+name: test flake8
+on: push
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.6", "3.7", "3.8", "3.9", "3.10", "3.11"]
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v4
+      with:
+          python-version: ${{ matrix.python-version }}
+    - name: Install dependencies
+      run: |
+          python -m pip install --upgrade pip
+          pip install flake8
+          pip install .
+    - name: instalation test
+      run: bidsme --help
+    - name: lint with flake8
+      run: |
+        flake8 .
