+++ .github/workflows/flake8.yml
@@ -3,6 +3,7 @@ on: push
 jobs:
   build:
     runs-on: ubuntu-latest
+    timeout-minutes: 10
     strategy:
       matrix:
         python-version: ["3.6.7", "3.7", "3.8", "3.9", "3.10", "3.11"]
