+++ .github/workflows/flake8.yml
@@ -2,11 +2,11 @@ name: test flake8
 on: push
 jobs:
   build:
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-18.04
     timeout-minutes: 10
     strategy:
       matrix:
-        python-version: ["3.6.7", "3.7", "3.8", "3.9", "3.10", "3.11"]
+        python-version: ["3.6", "3.7", "3.8", "3.9", "3.10", "3.11"]
 
     steps:
     - uses: actions/checkout@v3
@@ -18,7 +18,6 @@ jobs:
       run: |
           python -m pip install --upgrade pip
           pip install flake8
-          pip install .
     - name: instalation test
       run: bidsme --help
     - name: lint with flake8
