+++ .github/workflows/flake8.yml
@@ -17,6 +17,7 @@ jobs:
       run: |
           python -m pip install --upgrade pip
           pip install flake8
+          pip install .
     - name: instalation test
       run: bidsme --help
     - name: lint with flake8
