+++ .github/workflows/flake8.yml
@@ -2,11 +2,10 @@ name: test flake8
 on: push
 jobs:
   build:
-    runs-on: ubuntu-latest
-    timeout-minutes: 10
+    runs-on: ubuntu-18.04
     strategy:
       matrix:
-        python-version: ["3.7"]
+        python-version: ["3.6", "3.7", "3.8", "3.9", "3.10", "3.11"]
     steps:
     - uses: actions/checkout@v3
     - name: Set up Python ${{ matrix.python-version }}
