### 2023-08-07 a66ef91
Added path printing
