+++ .github/workflows/flake8.yml
@@ -27,8 +27,12 @@ jobs:
         cd
         git clone https://github.com/CyclotronResearchCentre/bidsme_examples.git
         cd bidsme_examples/example1
+        pwd
     - name: preparation
-      run: bidsme prepare source/ renamed/ -r nii=MRI --part-template resources/participants.json --plugin resources/plugins/rename_plugin.py
+      run: |
+        pwd
+        ls
+        bidsme prepare source/ renamed/ -r nii=MRI --part-template resources/participants.json --plugin resources/plugins/rename_plugin.py
     - name: mapping
       run:  bidsme map renamed/ bids/ --plugin resources/plugins/bidsify_plugin.py
     - name: process
