+++ .github/workflows/flake8.yml
@@ -27,15 +27,19 @@ jobs:
         cd
         git clone https://github.com/CyclotronResearchCentre/bidsme_examples.git
         cd bidsme_examples/example1
-        pwd
     - name: preparation
       run: |
-        pwd
-        ls
+        cd ~/bidsme_examples/example1
         bidsme prepare source/ renamed/ -r nii=MRI --part-template resources/participants.json --plugin resources/plugins/rename_plugin.py
     - name: mapping
-      run:  bidsme map renamed/ bids/ --plugin resources/plugins/bidsify_plugin.py
+      run:  |
+        cd ~/bidsme_examples/example1
+        bidsme map renamed/ bids/ --plugin resources/plugins/bidsify_plugin.py
     - name: process
-      run:  bidsme process renamed/ bids/ --plugin resources/plugins/process_plugin.py
+      run:  |
+        cd ~/bidsme_examples/example1
+        bidsme process renamed/ bids/ --plugin resources/plugins/process_plugin.py
     - name: bidsfify 
-      run: bidsme bidsify renamed/ bids/ --plugin resources/plugins/bidsify_plugin.py
+      run: |
+        cd ~/bidsme_examples/example1
+        bidsme bidsify renamed/ bids/ --plugin resources/plugins/bidsify_plugin.py
