+++ .github/workflows/flake8.yml
@@ -3,9 +3,10 @@ on: push
 jobs:
   build:
     runs-on: ubuntu-20.04
+
     strategy:
       matrix:
-        python-version: ["3.6", "3.7", "3.8", "3.9", "3.10", "3.11"]
+        python-version: ["3.6", "3.7", "3.8"]
     steps:
     - uses: actions/checkout@v3
     - name: Set up Python ${{ matrix.python-version }}
