+++ .github/workflows/flake8.yml
@@ -33,7 +33,7 @@ jobs:
         cd
         git clone https://github.com/CyclotronResearchCentre/bidsme_examples.git
         cd bidsme_examples/example1
-        pip install xldr
+        pip install openpyxl
     - name: preparation
       run: |
         cd ~/bidsme_examples/example1
