### 2023-08-07 6d46ece
Added paths to map
