+++ .github/workflows/flake8.yml
@@ -6,7 +6,7 @@ jobs:
 
     strategy:
       matrix:
-        python-version: ["3.6"]
+        python-version: ["3.6", "3.7", "3.8", "3.9", "3.10", "3.11"]
     steps:
     - uses: actions/checkout@v3
     - name: Set up Python ${{ matrix.python-version }}
