+++ .github/workflows/flake8.yml
@@ -16,7 +16,7 @@ jobs:
     - name: Install dependencies
       run: |
           python -m pip install --upgrade pip
-          pip install xlrd==1.2.0
+          pip install xlrd==1.2.0 openpyxl
           pip install flake8 wheel
           pip install .
     - name: instalation test
