+++ requirements.txt
@@ -1,4 +1,4 @@
-pandas
+pandas<=1.4.4
 ruamel.yaml>=0.15.35
 coloredlogs
 pydicom >= 1.4.2
