+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+
+### Added:
+ - Integrartion test for GitHub
+
 ## [1.4.1] - 2023-07-12
 
 ### Fixed:
