+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 
+## [1.4.2] - 2023-08-08
 ### Added:
  - Integrartion test for GitHub
 ### Changed:
