+++ joss_paper/paper.md
@@ -34,7 +34,7 @@ bibliography: paper.bib
 
 # Summary
 
-The purpose of Bidsme is to organize a given medical image dataset following the "Brain Image Dataset Structure" (BIDS) [@Gorgolewski2016]. Bidsme is an all-in-one organizer tool, that not only renames and re-structures the original data files, but also extracts and formats the necessary metadata. During the data organization, Bidsme provides the user with the full control over these processes, allowing the use of non-standard metadata and file names, as well as the addition of modalities not yet described by the BIDS. Instead of strictly imposing this structure, Bidsme allows the user to fully configure how the source dataset will be organized and what metadata will be included. Bidsme can be used both as python package and command-line tool, and includes a tutorial with a test dataset.
+The purpose of Bidsme is to organize a given medical image dataset following the "Brain Image Dataset Structure" (BIDS) [@Gorgolewski2016]. Bidsme is an all-in-one organizer tool, that not only renames and re-structures the original data files, but also extracts and formats the necessary metadata. During the data organization, Bidsme provides the user with the full control over these processes, allowing the use of non-standard metadata and file names, as well as the addition of modalities not yet described by the BIDS. Instead of strictly imposing this structure, Bidsme allows the user to fully configure how the source dataset will be organized and what metadata will be included. Bidsme can be used both as Python package and command-line tool, and includes a tutorial with a test dataset.
 
 # Statement of need
 
@@ -48,7 +48,7 @@ The challenge increases for datasets acquired prior to the introduction of BIDS,
 
 The ideal organizer tool must be able to be flexibly adapted to any original data structure and to any reasonable laboratory practices. It must try to retrieve as much necessary metadata as possible, but also allow the user to add additional metadata. It must suggest to follow the current standard but allow deviations from it, e.g. when a given modality is not defined in BIDS.
 
-These fundamental principles have been adopted as guidelines for the development of Bidsme. Bidsme gives full control of the data organization workflow to the user, imposing only the core of BIDS -- the folder structure, the file naming style and the global metadata. The actual names and set of entities are suggested to the user but are not imposed. Likewise, the user is free to add, remove or modify any automatically retrieved metadata.
+These fundamental principles have been adopted as guidelines for the development of Bidsme. Bidsme gives full control of the data organization workflow to the user, imposing only the core of BIDS -- the directory structure, the file naming style and the global metadata. The actual names and set of entities are suggested to the user but are not imposed. Likewise, the user is free to add, remove or modify any automatically retrieved metadata.
 
 # Bidsme overview and usage
 
@@ -58,13 +58,13 @@ The bidsification workflow using Bidsme is presented in \autoref{fig:workflow}. 
 ![Workflow of a bidsification using Bidsme. Dashed arrows and boxes represent optional steps.\label{fig:workflow}](plots/bidsme_schema.png)
 </center>
 
-The preparation step organizes the dataset into BIDS-like structure, with separate folders for each subject and session. The standardized structure of the prepared dataset not only facilitates the subsequent bidsification, but also helps with visual inspection of data integrity and provides an opportunity for intervention on the dataset (e.g. with removal of corrupted or failed data samples), while keeping the original dataset untouched. Several original datasets can be prepared into the same dataset, as long as there is no overlapping data. This can be useful when bidsifying datasets with several modalities (MRI, EEG, PET).
[…]
