### 2023-08-08 CyclotronResearchCentre/bidsme#12 @nbeliy
Hi @TheChymera,

I've added integration test, and fixed your smaller remarks. For the `global metadata` I still searching how to reformulate.
