### 2023-08-08 CyclotronResearchCentre/bidsme#7 @nbeliy
Hi @PeerHerholz,

I've put integration test (it was always there, but running on gitlab repository)
