### 2023-08-08 CyclotronResearchCentre/bidsme#11 @TheChymera
@nbeliy the `python-exec2` file is a generic wrapper, so it should be ok. Thanks.
