### 2023-08-08 CyclotronResearchCentre/bidsme#7 @PeerHerholz
Hi @nbeliy,

cool, looks great!
