### 2023-08-08 CyclotronResearchCentre/bidsme#12 @TheChymera
> Will it be better to remove of mention of global metadata completely?

Not necessarily, you can add a sentence saying “we define *global metadata* as ...”, then you write what it is, and “global metadata” will be a specific term introduced by your article. The problem is it sounds like it's a BIDS term, which it is not. I do think it would be better to not introduce a new term though, and just call it “select metadata”. Or if you only mention it once, just say what metadata you mean.

---

> Please consider splitting the source code over more lines

> Can you provide more details/examples?

I mean the “code” in the document source

I.e. instead of:

```
In this article we tackle this problem. This is an important problem because of the following implications. Our solution is amazing as demonstrated by the following results.
```

The following makes review easier. This is because I can reference line numbers instead of copying text that's otherwise hard to navigate to — also, it has the advantage over fixed-length line breaks, that you don't need to re-write the paragraph after you change the length of the first line :3

```
In this article we tackle this problem.
This is an important problem because of the following implications.
Our solution is amazing as demonstrated by the following results.
```

But it was just a suggestion, so don't worry. Maybe think of it for the next article, we can move on with the actual issues here. Please link commits with the fixes.
