### 2023-08-16 CyclotronResearchCentre/bidsme#12 @nbeliy
Hi @TheChymera ,

I rephrased the `core metadata` into :
> some minimal metadata (those included in `participants.tsv and` `*_scans.tsv` tabular files)

I believe that that fixes this issue (please correct if I missed something)
