### 2023-08-16 CyclotronResearchCentre/bidsme#10 @nbeliy
Forgot to announce in this issue, I ported the integration test from gitlab to github.
