### 2023-08-16 CyclotronResearchCentre/bidsme#9 @nbeliy
Hi @TheChymera ,

Updated to conventional schema (1.4.2)
