### 2023-08-16 CyclotronResearchCentre/bidsme#4 @nbeliy
Hi @PeerHerholz, sory for the long silence.

If I understand, the installation issue has been fixed on your side. In the readme, there a suggestion to use virtual envirement:
> It is recommended to use virtual environment when installing bidsme (more info [here](https://github.com/CyclotronResearchCentre/bidsme_tutorial#using-virtual-environments-and-kernels) and [here](https://docs.python.org/3/library/venv.html)).

Is it enough, or should I stress it more and with more instructions?
