### 2023-08-16 CyclotronResearchCentre/bidsme#4 @PeerHerholz
Hi @nbeliy,

no worries at all! I think it's good. A few thoughts I had: maybe rephrase `UNIX` to `venv` and outline that users only need the `ipykernel` step if they plan to use the respective `bidsme environment` in a `jupyter notebook`.

Sorry if this was already addressed somewhere but do you plan to add documentation in form of a webpage?
