### 2023-08-17 c715d88
Improved INSTALLATION.md
