+++ INSTALLATION.md
@@ -1,14 +1,6 @@
 ## Installation of bidsme
 
-Since version `1.4.0` bidsme can be installed using pip.
-It is the best option if you don't plan to contribute to `bidsme`
-developmnent.
-
-For advanced usage, and development, you can install bidsme
-manually (with or without pip).
-
-I will suggest to install `bidsme` within it's own virtual
-environment (see below).
+Since version `1.4.0` bidsme can be installed using pip. It is the best option if you don't plan to contribute to `bidsme` developmnent. For advanced usage, and development, you can install bidsme manually (with or without pip). I will suggest to install `bidsme` within it's own virtual environment (see below).
 
 ### Automatic installation using pip (not editable)
 
@@ -17,15 +9,14 @@ To install `bidsme` using `pip` you just need to write in terminal:
 pip install git+https://gitlab.uliege.be/CyclotronResearchCentre/Public/bidstools/bidsme/bidsme.git
 ```
 
-It should install `bidsme` and nessesary dependencies in the local
-python3 library.
+It should install `bidsme` and nessesary dependencies in the local Python3 library.
 
 Once installed, you should be able to run `bidsme` from terminal:
 ```bash
 bidsme --help
 ```
 
-or from `iPthon3`/`notebook`:
+or from Python console:
 ```python
 import bidsme
 bidsme.init()
@@ -39,34 +30,27 @@ cd <installation dir>
 git clone https://gitlab.uliege.be/CyclotronResearchCentre/Public/bidstools/bidsme/bidsme.git
 ```
 
-Once repository is cloned, you can add paths to python paths using `pip -e`:
+Once repository is cloned, you can add nessesary paths to Python paths using `pip -e`:
 ```bash
 pip install -e <installation dir>/bidsme
 ```
-In this case, pip will automatically install needed dependancies,
-also it will be aviable to be imported in python3 console or notebook:
+In this case, pip will automatically install needed dependancies, also it will be aviable to be imported in Python3 console or notebook:
 ```python
 import bidsme
 bidsme.init()
 ```
 
-Or just run bidsme directly:
+Or just run bidsme directly from cloned project:
 ```bash
 python3 <installation dir>/bidsme.py --help
 ```
-In this case you need to install the dependencies (see requirements.txt file).
+In this case you need to manually install the dependencies from [requirements.txt](requirements.txt) file.
 
 ## Using virtual environments and kernels
 
-`bidsme` will require the installation of some additional python packages, some of them
-are very common, like `pandas`, and likely already present in your installation of
-python, others are less common.
+`bidsme` will require the installation of some additional Python packages, some of them are very common, like `pandas`, and likely already present in your installation of Python, others are less common. In order to keep Python installation clean, usage of virtual environments and/or kernels are suggested.
 
-In order to keep python installation clean, usage of virtual environments and/or kernels
-are suggested.
-
-If you are using [virtual environment](https://docs.python.org/3/library/venv.html) and/or [(Ana)conda](https://anaconda.org/), then creating a new envoronment is
-straightforward, in terminal you just need to:
+If you are using [virtual environment](https://docs.python.org/3/library/venv.html) and/or [(Ana)conda](https://anaconda.org/), then creating a new envoronment is straightforward, in terminal you just need to:
 **[venv](https://docs.python.org/3/tutorial/venv.html)**:
 ``` 
 python3 -m venv bidsme_env
@@ -79,25 +63,25 @@ conda create --name bidsme_env
 conda activate bidsme_env
 ```
 
-In order to deactivate (return to your default) environment, you just need
-`deactivate` with venv or `conda deactivate` in conda.
+In order to deactivate (return to your default) environment, you just need `deactivate` with venv or `conda deactivate` in conda.
 
[…]
