### 2023-08-17 CyclotronResearchCentre/bidsme#4 @nbeliy
Hi @PeerHerholz,

I uploaded CONTRIBUTING, also added changes that you requested into INSTALLATION.

> Sorry if this was already addressed somewhere but do you plan to add documentation in form of a webpage?

Unfortunately no such plans exists -- it's one man project, and I don't really have time to create full web documentation(
