### 2023-08-17 ce0ec2d
Fixed bug with working with bare NIFTY files
