### 2023-08-17 CyclotronResearchCentre/bidsme#4 @PeerHerholz
Hi @nbeliy,

thanks implementing the changes and for the information! In that case, could you maybe add a little overview section to the `README.md`, ie linking to all respective parts (contributing, etc.)? Sorry for asking all those add-ons, I just think it might be easier for users to navigate.
