### 2023-08-23 CyclotronResearchCentre/bidsme#8 @nbeliy
Hi @PeerHerholz,

Can you tell if the current contribution guide is satisfactory for you? And if so, can you close the issue?
