### 2023-08-23 CyclotronResearchCentre/bidsme#4 @nbeliy
Hi @PeerHerholz ,

Is there something in this issue still unadressed, and if so, can you close it?

Thanks,
Nikita
