### 2023-08-23 CyclotronResearchCentre/bidsme#8 @PeerHerholz
Hi @nbeliy,

I think so, thanks for your work on this! I think the only thing left would be to mention/link it in the `README.md`.
