### 2023-08-23 CyclotronResearchCentre/bidsme#4 @PeerHerholz
Hi @nbeliy,

as mentioned in the prior comment, it would be nice to have a little overview with respective links in the `README.md` to make it easier for users to navigate.

Cheers, Peer
