### 2023-08-23 #5575 @emdupre
Hi @nbeliy, thanks for confirming ! I noticed that you hadn't yet commented on https://github.com/CyclotronResearchCentre/bidsme/issues/5 and https://github.com/CyclotronResearchCentre/bidsme/issues/6 ; could you confirm if these have already been addressed ?
