+++ doc/cli-interface.md
@@ -1,49 +1,74 @@
-## <a name="interface"></a> The BIDSme CLI interface
+# The bidsme CLI interface
 
-All interactions with BIDSme occurs from command-line interface, by a master script [`bidsme.py`](https://gitlab.uliege.be/CyclotronResearchCentre/Public/bidstools/bidsme/bidsme/-/blob/tutorial/bidsme/bidsme.py).
 
-This script accepts a small set of parameters and commands 
 
-- `prepare`, to prepare dataset
-- `process`, to process dataset
-- `bidsify`, to bidsify dataset
-- `map`, to create bidsmap
+If the installation of the bidsme was performed using `pip` (see [instructions](../INSTALLATION.md)), then the command line interface `bidsme` and `bidsme-pdb` should be available in the terminal.
+The two interfaces are identical except the second one executes within a standard Python3 debugger [pdb](https://docs.python.org/3/library/pdb.html).
 
-Outside the standard `-h`, `-v` options, to access help and retrieve the version of the toolkit, `bidsme` accepts 
-the option `-c', '--configuration` that retrieves the path to the configuration file. 
-This file is searched following the precedence order: current directory, user-standard directory and 
-bids code directory (when available).
+If `bidsme` was installed manually, then the same interface can be achieved by calling `main.py:main` function.
 
-The `--conf-save` switch saves the current configuration (affected by command-line
-options) in the given location. It is useful to run this option once to update the configuration.
+## The main interface
 
-> N.B. both `-c` and `--conf-save` must be given **before** the command 'prepare', 'process' or 'bidsify'.
+```bash
+usage: bidsme [-h] [-c conf.yaml] [--conf-save] [-v]  subcommand
+```
 
-The individual commands accept common and individual arguments. 
-In what follows, only common arguments are described, whereby individual ones are 
-described in corresponding sections.
+The main interface `bidsme` require a sub-command, corresponding to one of the bidsification step:
 
-- Logging options, corresponding to the *logging* section of configuration file:
-    *  `-q`, `--quiet` suppress the standard output, useful for running in the script
-    * `--level` sets the message verbosity of the log output, from very verbose *DEBUG*
-    to showing only critical message *CRITICAL*
-    * `--formatter` sets the log line format message 
-- Plug-in options, corresponding to the *plugins* section of configuration file. They affect only the
+ - `prepare`, to prepare dataset
+ - `process`, to process dataset
+ - `bidsify`, to bidsify dataset
+ - `map`, to create bidsmap
+
+For example,
+```bash
+bidsme bidsify <options>
+```
+will execute bidsification step.
+
+Some general optional parameter can be passed **before** the command:
+  
+ - `-h` prints help for main interface
+ - `-v` prints the version of program and version of supported BIDS
+ - `-c`, '--configuration` with path to the configuration yaml file allow to retrieve the saved parameters
+ - `--conf-save` will save current command-line parameters into configuration file 
+
+## Sub-command interface
+
+All sub-commands will require two positional arguments:
+
+ - `source` -- path to dataset on which subcommand must be executed
+ - `destination` -- path where the results of the subcommand will be written
+
+For the `prepare` subcommand, source is the raw dataset and destination will be prepared dataset.
+For the `process` subcommand, source and destination is prepared dataset.
+For all other subcommands the source is prepared dataset, and destination is bidsified dataset.
+
+All subcommands will have common parameters, described below.
+
+### Logging options
+Logging options, corresponding to the *logging* section of configuration file:
+ *  `-q`, `--quiet` suppress the standard output, useful for running in the script
+ * `--level` sets the message verbosity of the log output, from very verbose *DEBUG*, to showing only critical message *CRITICAL*
[…]
