+++ doc/creating_map.md
@@ -1,14 +1,17 @@
-This is how I usually do the mapping, step-by-step. Sometimes I do some shortcuts, but I suggests to avoid them until you feel comfortable.
+# Bidsmap creation
+
+The bidsmap creation is the most annoying and complex part of bidsification with `bidsme`.
+In this instruction we tried to provide the step-by-step instructions how to create the map.
 
 
-# Intro and definitions
+## Intro and definitions
 
 There main guidelines:
 
   - Have a list of series per session in their expected order, useful to retrieve names and make notes
   - Prepare bidsified names beforehand, this will streamline process and avoid errors in `IntendedFor` meta-fields
   - Work on a good subject i.e. complete without deviation from protocol, without duplicated acquisition etc.
-  - Work at 1 warning/error/recording at time, to remember modifications in previous step
+  - `bidme map` will stop at each error and warning, so make sure to fix errors and warnings that `bidsme` report
   - Rerun map command at each occasion
 
 Some vocabulary (just to avoid confusion):
@@ -21,11 +24,11 @@ Some vocabulary (just to avoid confusion):
   - data type -- kind of data acquired (MRI, EEG, PET etc.)
   - data format -- format that recordings are stored (dicom, nifti, nifti+json etc.)
   - modality -- type of recording, corresponds to lowest level folder in bids structure (func, anat, dwi, beh, edf etc.)
-  - suffix -- BIDS identificant differentiating different types of recording within modality, last part of bidsified name
-  - entity -- a underscore-separated blocks constructing bidsified name, entity is composed of name and value, separated by dash. entity must be as uniform as possible within modality, values must be intuitive
+  - suffix -- BIDS identifier differentiating different types of recording within modality, last part of bidsified name
+  - entity -- a underscore-separated blocks constructing bidsified name, entity is composed of name and value, separated by dash. Entity must be as uniform as possible within modality, values must be intuitive
   - map entry -- bidsmap block corresponding to one recording, starting with `provenance` field and ending with `json`
 
-For MRI the relevant page is (there)[https://bids-specification.readthedocs.io/en/v1.2.0/04-modality-specific-files/01-magnetic-resonance-imaging-data.html]. The current proposal for MPM is (there)[https://docs.google.com/document/d/1QwfHyBzOyFWOLO4u_kkojLpUhW0-4_M7Ubafu9Gf4Gg/edit].
+For MRI the relevant page is [there](https://bids-specification.readthedocs.io/en/v1.2.0/04-modality-specific-files/01-magnetic-resonance-imaging-data.html).
 
 In following I will use `[]` to indicate context-depending variables, usually specific paths.
 For example `[path-to-source-dataset]` should be replaced by by your specific path to source dataset.
@@ -33,86 +36,138 @@ Don't worry about plugins at this stage, but if in your map you find something n
 
 If you are not sure on name to give to one entity, or suffix, just use something easy to spot, like "ToRename".
 
+
+### Prepared data
+
 In the prepared folder data is always organized as follows:
 ```
 sub-[subject Id]/ses-[session Id]/[data type]/[series number]-[series Id]/data
 ```
 
-The bidsme log output usually cites concerning series, if not look to the line above for currently scanned series. Sometimes you need to look into the header, use the organisation to easily find correct file.
+### Errors/warning logs
 
+`bidsme` will try to provide as much of information about encountered errors and warnings as possible. Read the printed log not only at error, but also before -- `bidsme` will say on which series he worked when error occurred. 
 
-Use aliases to invoke bidsme, for ex.:
-```
[…]
