### 2023-08-25 CyclotronResearchCentre/bidsme#8 @nbeliy
Hi @PeerHerholz ,
I updated readme, can you check if references are clearer now?
