### 2023-08-25 CyclotronResearchCentre/bidsme#6 @nbeliy
The example is the tutorial. I have plans to make more direct example, but it may take time...
