### 2023-08-25 CyclotronResearchCentre/bidsme#4 @nbeliy
Can you check if new readme is better now?
