### 2023-08-25 CyclotronResearchCentre/bidsme#5 @nbeliy
I've updated the documentation in doc, but not sure if will ever have time to create stand-alone documentation with the API.
