### 2023-08-25 CyclotronResearchCentre/bidsme#5 @PeerHerholz
Hi @nbeliy,

thanks for the update. Just confirming with @emdupre if this works re the `JOSS` requirements?
