### 2023-08-25 CyclotronResearchCentre/bidsme#4 @PeerHerholz
Hi @nbeliy,

yes, thanks for implementing the changes!
