### 2023-08-28 CyclotronResearchCentre/bidsme#5 @emdupre
Hi @PeerHerholz and @nbeliy,

Thanks for starting this conversation ! To confirm, the JOSS review criteria for API documentation are [linked here](https://joss.readthedocs.io/en/latest/review_criteria.html#api-documentation) and note:

> Reviewers should check that the software API is documented to a suitable level.
>     **Good:** All functions/methods are documented including example inputs and outputs
>     **OK:** Core API functionality is documented
>     **Bad (not acceptable):** API is undocumented

To confirm, then : we require---at a minimum---that all of the core API functionality is documented. It seems that the API documentation has been manually generated to date in the `doc` folder. This makes it somewhat difficult to confirm if all core functionality is covered, which is likely where the confusion arose.

@PeerHerholz and @TheChymera, when you re-review the docs, would you be able to read through with this consideration in mind ?

Otherwise @nbeliy, as @PeerHerholz suggested, you could use Sphinx to auto-generate all of your API documentation, such that it would automatically reach **Good** status per JOSS's definitions. This is not required for the JOSS review process, though !
