### 2023-08-29 CyclotronResearchCentre/bidsme#6 @emdupre
Hi @nbeliy and @PeerHerholz, and thanks for starting this conversation !

For the JOSS review process, we require:

> Example usage
> The authors should include examples of how to use the software (ideally to solve real-world analysis problems).

as [noted here](https://joss.readthedocs.io/en/latest/review_criteria.html#example-usage). The tutorial content appears to meet this requirement, though I would like to hear @PeerHerholz and @TheChymera 's opinions.

I notice, though, that the tutorial is only included in another repository (i.e., `bidsme_tutorial`). Since `bidsme` rather than `bidsme_tutorial` has been submitted for review, I would expect that the tutorial would be included in this code base. Could you please correct this @nbeliy, such that only code in the current repository is under review ?
