### 2023-08-29 CyclotronResearchCentre/bidsme#5 @ChristophePhillips
Hi @emdupre and @PeerHerholz,

Thank you for your constructive feedback!

FYI @nbeliy is currently enjoying a well deserved holiday break, so we'll come back to you in a couple of weeks only. 😃
