### 2023-09-06 CyclotronResearchCentre/bidsme#12 @TheChymera
Yes, that addresses the issue.
