### 2023-09-06 CyclotronResearchCentre/bidsme#10 @TheChymera
The test suite appears to pass on GitHub CI, however I can't seem to run it locally, am I missing something or is there an issue with the documentation of dependencies and testdata?

```
[deco]~/src/bidsme/bidsme_examples/example1 ❱ bidsme process renamed/ bids/ --plugin resources/plugins/process_plugin.py -b resources/map/bidsmap.yaml
main(76) - INFO
main(77) - INFO -------------- START bidsme ----------------
main(78) - INFO Wed Sep  6 04:21:56 2023
main(79) - INFO programm version: 1.4.0.post4
main(80) - INFO bids version: 1.2.0
bidsme.process(189) - INFO -------------- Processing data -------------
bidsme.process(190) - INFO Source directory: renamed/
bidsme.process(191) - INFO Destination directory: bids/
bidsme.process(213) - WARNING Dataset description file 'dataset_description.json' not found in 'bids/'
bidsme.process(219) - WARNING Dataset readme file 'README' not found in 'bids/'
bidsme.process(234) - INFO loading bidsmap /home/chymera/src/bidsme/bidsme_examples/example1/resources/map/bidsmap.yaml
bidsme.plugins.plugins(79) - INFO Loading module process_plugin from resources/plugins/process_plugin.py
bidsme.exceptions.exceptions(49) - ERROR /usr/lib/python3.11/site-packages/bidsme/main.py(138) in main:
bidsme.exceptions.exceptions(49) - ERROR /usr/lib/python3.11/site-packages/bidsme/process.py(265) in process:
bidsme.exceptions.exceptions(49) - ERROR /usr/lib/python3.11/site-packages/bidsme/bidsMeta/BidsSession.py(213) in loadSubjectFields:
bidsme.exceptions.exceptions(49) - ERROR /usr/lib/python3.11/site-packages/bidsme/bidsMeta/BidsMeta.py(435) in LoadDefinitions:
bidsme.exceptions.exceptions(51) - ERROR 1:FileNotFoundError: [Errno 2] No such file or directory: 'renamed/participants.json'
main(181) - INFO Command: ['process', 'renamed/', 'bids/', '--plugin', 'resources/plugins/process_plugin.py', '-b', 'resources/map/bidsmap.yaml']
main(183) - INFO -------------- FINISHED! -------------------
main(148) - INFO counthangler:{'WARNING': 2, 'ERROR': 5}
main(152) - INFO loghandler:/home/chymera/src/bidsme/bidsme_examples/example1/bids/code/bidsme/process/main.log
main(152) - INFO errorhandler:/home/chymera/src/bidsme/bidsme_examples/example1/bids/code/bidsme/process/main.err
main(185) - INFO Took 0.891542123 seconds
main(186) - INFO --------------------------------------------
[deco]~/src/bidsme/bidsme_examples/example1 ❱ bidsme bidsify renamed/ bids/ --plugin resources/plugins/bidsify_plugin.py -b resources/map/bidsmap.yaml
main(76) - INFO
main(77) - INFO -------------- START bidsme ----------------
main(78) - INFO Wed Sep  6 04:22:09 2023
main(79) - INFO programm version: 1.4.0.post4
main(80) - INFO bids version: 1.2.0
bidsme.bidsify(179) - INFO -------------- Prepearing data -------------
bidsme.bidsify(180) - INFO Source directory: renamed/
bidsme.bidsify(181) - INFO Destination directory: bids/
bidsme.bidsify(203) - WARNING Dataset description file 'dataset_description.json' not found in 'bids/'
bidsme.bidsify(209) - WARNING Dataset readme file 'README' not found in 'bids/'
bidsme.bidsify(224) - INFO loading bidsmap /home/chymera/src/bidsme/bidsme_examples/example1/resources/map/bidsmap.yaml
bidsme.plugins.plugins(79) - INFO Loading module bidsify_plugin from resources/plugins/bidsify_plugin.py
bidsme.exceptions.exceptions(49) - ERROR /usr/lib/python3.11/site-packages/bidsme/main.py(151) in main:
bidsme.exceptions.exceptions(49) - ERROR /usr/lib/python3.11/site-packages/bidsme/bidsify.py(255) in bidsify:
bidsme.exceptions.exceptions(49) - ERROR /usr/lib/python3.11/site-packages/bidsme/bidsMeta/BidsSession.py(213) in loadSubjectFields:
bidsme.exceptions.exceptions(49) - ERROR /usr/lib/python3.11/site-packages/bidsme/bidsMeta/BidsMeta.py(435) in LoadDefinitions:
bidsme.exceptions.exceptions(51) - ERROR 1:FileNotFoundError: [Errno 2] No such file or directory: 'renamed/participants.json'
main(181) - INFO Command: ['bidsify', 'renamed/', 'bids/', '--plugin', 'resources/plugins/bidsify_plugin.py', '-b', 'resources/map/bidsmap.yaml']
main(183) - INFO -------------- FINISHED! -------------------
main(148) - INFO counthangler:{'WARNING': 2, 'ERROR': 5}
main(152) - INFO loghandler:/home/chymera/src/bidsme/bidsme_examples/example1/bids/code/bidsme/bidsify/main.log
main(152) - INFO errorhandler:/home/chymera/src/bidsme/bidsme_examples/example1/bids/code/bidsme/bidsify/main.err
main(185) - INFO Took 0.973882708 seconds
main(186) - INFO --------------------------------------------
[deco]~/src/bidsme/bidsme_examples/example1 ❱ bidsme map renamed/ bids/ --plugin resources/plugins/bidsify_plugin.py -b resources/map/bidsmap.yaml
main(76) - INFO
main(77) - INFO -------------- START bidsme ----------------
main(78) - INFO Wed Sep  6 04:22:20 2023
main(79) - INFO programm version: 1.4.0.post4
main(80) - INFO bids version: 1.2.0
bidsme.mapper(202) - INFO ------------ Generating bidsmap ------------
bidsme.mapper(203) - INFO Current directory: /home/chymera/src/bidsme/bidsme_examples/example1
bidsme.mapper(204) - INFO Source directory: renamed/
bidsme.mapper(205) - INFO Destination directory: bids/
bidsme.mapper(221) - INFO loading template bidsmap bidsmap_template.yaml
bidsme.bidsmap._bidsmap(103) - WARNING Failed to find type EEG/BrainVision readed from /usr/lib/python3.11/site-packages/bidsme/heuristics/bidsmap_template.yaml
bidsme.mapper(240) - INFO loading working bidsmap /home/chymera/src/bidsme/bidsme_examples/example1/resources/map/bidsmap.yaml
bidsme.bidsmap._bidsmap(64) - INFO bids/code/bidsme/unknown.yaml not found. Bidsmap will be empty
bidsme.plugins.plugins(79) - INFO Loading module bidsify_plugin from resources/plugins/bidsify_plugin.py
bidsme.bidsMeta.BidsTable(88) - ERROR participants.tsv: Unable to find definitions file
bidsme.exceptions.exceptions(49) - ERROR /usr/lib/python3.11/site-packages/bidsme/main.py(164) in main:
bidsme.exceptions.exceptions(49) - ERROR /usr/lib/python3.11/site-packages/bidsme/mapper.py(265) in mapper:
bidsme.exceptions.exceptions(49) - ERROR /usr/lib/python3.11/site-packages/bidsme/bidsMeta/BidsTable.py(90) in __init__:
bidsme.exceptions.exceptions(51) - ERROR 1:FileNotFoundError: renamed/participants.json
main(181) - INFO Command: ['map', 'renamed/', 'bids/', '--plugin', 'resources/plugins/bidsify_plugin.py', '-b', 'resources/map/bidsmap.yaml']
main(183) - INFO -------------- FINISHED! -------------------
main(148) - INFO counthangler:{'WARNING': 1, 'ERROR': 5}
main(152) - INFO loghandler:/home/chymera/src/bidsme/bidsme_examples/example1/bids/code/bidsme/map/main.log
main(152) - INFO errorhandler:/home/chymera/src/bidsme/bidsme_examples/example1/bids/code/bidsme/map/main.err
main(185) - INFO Took 1.076100048 seconds
main(186) - INFO --------------------------------------------
[deco]~/src/bidsme/bidsme_examples/example1 ❱ git rev-parse HEAD
8c7f74742fc4b2db92230160cbb98185f5ab2920
```
