### 2023-09-06 CyclotronResearchCentre/bidsme#9 @TheChymera
That solves the issue.
