### 2023-09-06 CyclotronResearchCentre/bidsme#10 @TheChymera
Also, I would strongly recommend the test suite be captured in a dedicated script or [BATS](https://github.com/bats-core/bats-core) (since you're calling the package CLI via Bash), or in a `test` module as used by `pytest` (if you want the Python functions tested directly in Python) so that they can be run automatically without copy-pasting from the GitHub test suite. In the GitHub test suite, you can then call that script/BATS/pytest.

The reason why this is important is that package managers might want to integrate the testing functionality of your package so users can automatically ascertain that what they installed works as it should on their machine. Currently this is not really possible.
