### 2023-09-06 CyclotronResearchCentre/bidsme#6 @TheChymera
@emdupre I believe examples have been adequately provided, in the tutorials as well as the test suite the author has now set up. As for tutorials being included in the actual repository or elsewhere, I guess ultimately it's up to editorial discretion — I'd just mention that there is such a thing as bundling too many things together, and perhaps users of the package don't want to download the tutorial source alongside with the code.
