### 2023-09-06 CyclotronResearchCentre/bidsme#6 @emdupre
Thanks for confirming, @TheChymera !

Yes, I don't mean to suggest that everything associated with BIDSme should be in one place ; we do expect, though that there is sufficient documentation inside the repository under review ! If there is other content already in the repository that meets that the JOSS example usage criteria (noted above) this would also be fine. It seems that @PeerHerholz was not originally able to find it, though this might have changed as the test suite was added. I'll let you (@TheChymera and @PeerHerholz) confirm.
