### 2023-09-07 CyclotronResearchCentre/bidsme#6 @PeerHerholz
Hi folks,

thanks for following up on this. My comment was more related to the structure and outline of the examples.
I think with the recent changes and current state [here](https://github.com/CyclotronResearchCentre/bidsme/tree/joss_paper) potential users can find and navigate examples/tutorials sufficiently. Thus, if everyone is ok with the current state, I'm too.
