### 2023-09-11 #5575 @emdupre
Hi everyone, and happy Monday ! 🌻

Just following up from my previous message : these remaining issues were subsequently discussed in their individual threads.

@TheChymera and @PeerHerholz, thank you again for your feedback on bidsme ! If you could please revisit your individual reviewer checklists now that @nbeliy has finished addressing your initial round of comments, that will help us to move the submission forward !
