### 2023-09-11 CyclotronResearchCentre/bidsme#10 @nbeliy
Hi @TheChymera,

About the failing test: I'm not sure but it looks like the `bidsme process` is run not after `bidsme prepare` (maybe you run test in parallel or you remove the artifacts after the `prepare`). It is an integration test, not unit test (unfortunately).

For the bats, thanks for pointing it, I didn't knew about. But here it looks like an overkill -- just to replace one line of script in workflow.
