### 2023-09-15 CyclotronResearchCentre/bidsme#5 @nbeliy
Hi @emdupre and @PeerHerholz,
I will try to add documentation on readthesedocs, but it may take some time. I used human-readable python docstrings, and importing them using doxigen or sphynx will require rewriting all docstrings.

I tried to use pydoc, and it worked quite nice, but not sure if it's compatible with readthesedocs.
