### 2023-09-28 #5575 @emdupre
👋 Hi @TheChymera, I just wanted to quickly follow-up to confirm if you've been able to review your review checklist in light of the recent changes to _Bidsme_. If you have any questions or concerns, please don't hesitate to let me know.
