### 2023-10-20 #5575 @emdupre
Just following up to note that I've also reached out to @TheChymera over email to check on the status of this. I've asked for confirmation within the next week that they will be able to finish this review.

Thanks for your patience, @nbeliy, and thank you @TheChymera and @PeerHerholz for your engagement in reviewing _Bidsme_ to date !
