### 2023-10-23 CyclotronResearchCentre/bidsme#10 @TheChymera
Ok, the tests now pass locally. The environment needed openpyxl to be installed, which `pip` did not auto-install for some reason.
