### 2023-10-24 #5575 @emdupre
A few editorial comments :

**Major comments**
- In the software paper _Statement of Need_, it would be useful to directly explain how `bidsme` relates to existing other BIDS-formatting tools such as [Heudiconv](https://heudiconv.readthedocs.io/en/latest/) and [dcm2bids](https://unfmontreal.github.io/Dcm2Bids/3.1.1/). This falls under the "State of the field" review criteria.

**Minor comments**
- Please capitalize BIDS when it appears as an acronym, outside of Bidsme. See L51, L62, etc.
- L7, L22: Please note that the correct name for BIDS is "Brain Imaging Data Structure" rather than "Brain Image Data Structure". Please update this throughout.
- L28: Please change ", i.e." should be updated to "; i.e.,". This update also holds for other abbreviations like "e.g." on L33, L37, L42,
- L74. Please standardize "plug-in" to "plugin" as used on L75, or vice versa. Regardless, please review the text to ensure consistent spelling of this term.
- L99. "developed basing on" --> "developed based on"
- In Fig 1., the text under "source, "prepared," and "bidisifed" datasets listing the directory structures is difficult to read. Could you please increase the font size, likely by showing fewer files ?
