### 2023-11-14 #5575 @emdupre
Hi @nbeliy, I'm back in office and just wanted to check in on this ! Have you had a chance to integrate the recommended changes for the software paper ?
