### 2023-11-14 #5575 @nbeliy
Hi @emdupre ,

I did the smaller changes but got quite charged last weeks to compare tools. I'll regenerate the draft, can you check if the figure is readable now?
