### 2023-11-15 5f3c808
Added bib entries for dcm2bids and heudiconv
