### 2023-11-15 e3173cb
First draft for dcm2bids
