+++ joss_paper/paper.bib
@@ -54,3 +54,13 @@ ABSTRACT={Analyses of brain function and anatomy using shared neuroimaging data 
   version = {1.0.0},
   year = {2023}
 }
+
+@software{PET2BIDS_2022,
+  author = {Galassi, Anthony and Norgaard, Martin and Thomas, Adam G. and Gonzalez-Escamilla, Gabriel and Svarer, Claus and Rorden, Chris and Matheson, Granville and Knudsen, Gitte and Innis, Robert and Ganz-Benjaminsen, Melanie and Pernet, Cyril},
+  license = {MIT},
+  month = oct,
+  title = {{PET2BIDS}},
+  url = {https://github.com/openneuropet/PET2BIDS},
+  version = {1.0.0},
+  year = {2022}
+}
