### 2023-11-15 50311a4
Improved critique on dcm2bids
