### 2023-11-15 ab20e27
Changed encode to retrieve
