### 2023-11-15 #5575 @emdupre
Hi @nbeliy, the figure does look more legible, thank you ! Please let me know when you've finished making the other requested changes and I can continue processing the submission.

Also just to note: Editorial Bot only responds if the command is at the start of a GitHub comment. Apologies for any confusion !
