### 2023-11-17 bd60797
Added weird character to autors name
