### 2023-11-17 adf4fec
Concurrence (#13)
* Added bib entries for dcm2bids and heudiconv

* First draft for dcm2bids

* Added pet2bids reference

* Improved critique on dcm2bids

* Changed encode to retrieve

* Arianna's comments

* Added weird character to autors name

---------

Co-authored-by: Beliy Nikita <beliy.nikita@uliege.be>
