+++ joss_paper/paper.bib
@@ -35,3 +35,32 @@ ABSTRACT={Analyses of brain function and anatomy using shared neuroimaging data 
   journal = {Journal of Neuroscience Methods}
 }
 
+@software{BoreDcm2Bids2023,
+  author = {Boré, Arnaud and Guay, Samuel and Bedetti, Christophe and Meisler, Steven and GuenTher, Nick},
+  doi = {10.5281/zenodo.8342572},
+  month = aug,
+  title = {{Dcm2Bids}},
+  url = {https://github.com/UNFmontreal/Dcm2Bids},
+  version = {3.1.0},
+  year = {2023}
+}
+
+@software{heudiconv23,
+  author = {Halchenko, Yaroslav and Goncalves, Mathias and Velasco, Pablo and di Oleggio Castello, Matteo Visconti and Ghosh, Satrajit and Salo, Taylor and Wodder II, John T. and Hanke, Michael and Sadil, Patrick and Christian, Horea and Dae, Michael and Tilley, Steven and Kent, James and T, Isaac and Brett, Matthew and Amlien, Inge and Gorgolewski, Chris and Markiewicz, Chris and Lukas, Darren Christopher and Callenberg, Keith and Kahn, Aksoo Ari and Macdonald, Austin and Poldrack, Benjamin and Melo, Bruno and Braun, Henry and Lee, John and Pellman, John and Michael},
+  doi = {10.5281/zenodo.8364586},
+  month = sep,
+  title = {{nipy/heudiconv}},
+  url = {https://heudiconv.readthedocs.io/en/latest/index.html},
+  version = {1.0.0},
+  year = {2023}
+}
+
+@software{PET2BIDS_2022,
+  author = {Galassi, Anthony and Norgaard, Martin and Thomas, Adam G. and Gonzalez-Escamilla, Gabriel and Svarer, Claus and Rorden, Chris and Matheson, Granville and Knudsen, Gitte and Innis, Robert and Ganz-Benjaminsen, Melanie and Pernet, Cyril},
+  license = {MIT},
+  month = oct,
+  title = {{PET2BIDS}},
+  url = {https://github.com/openneuropet/PET2BIDS},
+  version = {1.0.0},
+  year = {2022}
+}
