### 2023-11-17 #5575 @nbeliy
@emdupre

> **Major comments**
>
>     * In the software paper _Statement of Need_, it would be useful to directly explain how `bidsme` relates to existing other BIDS-formatting tools such as [Heudiconv](https://heudiconv.readthedocs.io/en/latest/) and [dcm2bids](https://unfmontreal.github.io/Dcm2Bids/3.1.1/). This falls under the "State of the field" review criteria.

 - Done, added reference to 3 tools (2 that you mentioned and one for PET). It's not direct comparison but statement of issues that `bidsme` addresses.  I think it suits more the flow of paper. But if direct comparison is needed, I can rearrange these paragraphs.

> **Minor comments**
>     * Please capitalize BIDS when it appears as an acronym, outside of Bidsme. See L51, L62, etc.

 -  I capitalized BIDS, but should I also capitalize `bidsification`? BIDSification looks ugly for me, but I don't know the rules for derivative words from acronyms.

>     * L7, L22: Please note that the correct name for BIDS is "Brain Imaging Data Structure" rather than "Brain Image Data Structure". Please update this throughout.

 - done

>     * L28: Please change ", i.e." should be updated to "; i.e.,". This update also holds for other abbreviations like "e.g." on L33, L37, L42,

- done. Just for myself, this comma is purely stylistic? I often seen `i.e.` without comma.

>     * L74. Please standardize "plug-in" to "plugin" as used on L75, or vice versa. Regardless, please review the text to ensure consistent spelling of this term.

 - done.
>     * L99. "developed basing on" --> "developed based on"

-  done
>     * In Fig 1., the text under "source, "prepared," and "bidisifed" datasets listing the directory structures is difficult to read. Could you please increase the font size, likely by showing fewer files ?

- done
