### 2023-11-22 #5575 @emdupre
Thank you, @nbeliy !

> I capitalized BIDS, but should I also capitalize bidsification? BIDSification looks ugly for me, but I don't know the rules for derivative words from acronyms.

I would make these more explicit ; for example "BIDS-ificiation" and "BIDS-ify." This is also more likely to avoid confusion e.g., with the [`bidsify` tool](https://github.com/NILAB-UvA/bidsify).

> Just for myself, this comma is purely stylistic? I often seen i.e. without comma.

This is the correct in [American style guides](https://jakubmarian.com/comma-after-i-e-and-e-g/), which JOSS follows.
