### 2023-11-22 #5575 @emdupre
After making these changes, could you then please:

- [ ] Archive the reviewed software in [Zenodo](https://zenodo.org/) or a similar service (e.g., figshare, an institutional repository)
- [ ] Check the archival deposit (e.g., in Zenodo) has the correct metadata (e.g., the JOSS paper and software archive have the same title and author listing)
- [ ] List the DOI of the archived version here.
