### 2023-11-22 8242487
Capitalized BIDS everywere
