+++ joss_paper/paper.md
@@ -36,13 +36,13 @@ bibliography: paper.bib
 
 # Summary
 
-The purpose of Bidsme is to organize a given medical image dataset following the "Brain Imaging Data Structure" (BIDS) [@Gorgolewski2016]. Bidsme is an all-in-one organizer tool, that not only renames and re-structures the original data files, but also extracts and formats the necessary metadata. During the data organization, Bidsme provides the user with the full control over these processes, allowing the use of non-standard metadata and file names, as well as the addition of modalities not yet described by the BIDS. Instead of strictly imposing this structure, Bidsme allows the user to fully configure how the source dataset will be organized and what metadata will be included. Bidsme can be used both as Python package and command-line tool, and includes a tutorial with a test dataset.
+The purpose of Bidsme is to organize a given medical image dataset following the "Brain Imaging Data Structure" [BIDS; @Gorgolewski2016]. Bidsme is an all-in-one organizer tool, that not only renames and re-structures the original data files, but also extracts and formats the necessary metadata. During the data organization, Bidsme provides the user with the full control over these processes, allowing the use of non-standard metadata and file names, as well as the addition of modalities not yet described by the BIDS. Instead of strictly imposing this structure, Bidsme allows the user to fully configure how the source dataset will be organized and what metadata will be included. Bidsme can be used both as Python package and command-line tool, and includes a tutorial with a test dataset.
 
 # Statement of need
 
 For a long time, the neuroimaging community suffered from a lack of standardized file structure, formats and metadata conventions. Different laboratories, and even different research groups within a laboratory, had their own, idiosyncratic ways to organize their data. Analyses were then performed with scripts tailored to specific data structures, which made analysis workflow and data sharing unnecessarily complicated. Consequently, it was difficult to ensure results repeatability and validation.
 
-The Brain Imaging Data Structure (BIDS) [@Gorgolewski2016] was introduced to change this situation; it imposes a standard data structure and defines the required associated metadata. Once a dataset follows the imposed structure, any analysis tool, supporting the BIDS, should be able to automatically find the needed data to process it.
+The Brain Imaging Data Structure [BIDS; @Gorgolewski2016] was introduced to change this situation; it imposes a standard data structure and defines the required associated metadata. Once a dataset follows the imposed structure, any analysis tool, supporting the BIDS, should be able to automatically find the needed data to process it.
 
 Prior to the introduction of the BIDS, the main challenge was to adapt the processing scripts to the different (typically inconsistent) dataset structures. With the advent of the BIDS, the main challenge is now to "BIDS-ify" a given dataset, i.e., adapt a dataset to the BIDS. Not only the image files must be renamed according to the standard, but also the associated metadata must contain all expected values, using expected conventions and measurement units.
 
