+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [1.5.0] - 2023-11-22
+
+Release for the publication in JOSS
+
+### Fixed
+ - Module/base: Bug when  `recIdentity()` wasn't called as function
+
+### Changed
+ - Several improvements in documentation and README
+ - Updated links to the examples and tutorial to point to GitHub
+
 ## [1.4.3] - 2023-08-17
 ### Fixed:
  - Bug in bidsification when working with bare nifty file
