### 2023-11-22 238ebe5
Cleaned up JOSS-related files
