+++ joss_paper/paper.bib
@@ -1,66 +1,0 @@
-@article{Gorgolewski2016,
-  title={The brain imaging data structure, a format for organizing and describing outputs of neuroimaging experiments},
-  author={Gorgolewski, Krzysztof J. and Auer, Tibor and Calhoun, Vince D. and Craddock, R. Cameron and Das, Samir and Duff, Eugene P. and Flandin, Guillaume and Ghosh, Satrajit S. and Glatard, Tristan and Halchenko, Yaroslav O. and Handwerker, Daniel A. and Hanke, Michael and Keator, David and Li, Xiangrui and Michael, Zachary and Maumet, Camille and Nichols, B. Nolan and Nichols, Thomas E. and Pellman, John and Poline, Jean-Baptiste and Rokem, Ariel and Schaefer, Gunnar and Sochat, Vanessa and Triplett, William and Turner, Jessica A. and Varoquaux, Gaël and Poldrack, Russell A.},
-  journal={Scientific Data},
-  volume={3},
-  number={160044},
-  pages={},
-  year={2016},
-  doi={10.1038/sdata.2016.44},
-  publisher={Nature}
-}
-
-@article{Zwiers2022,
-AUTHOR={Zwiers, Marcel Peter and Moia, Stefano and Oostenveld, Robert},
-TITLE={BIDScoin: A User-Friendly Application to Convert Source Data to Brain Imaging Data Structure},
-JOURNAL={Frontiers in Neuroinformatics},
-VOLUME={15},
-YEAR={2022},
-URL={https://www.frontiersin.org/articles/10.3389/fninf.2021.770608},
-DOI={10.3389/fninf.2021.770608},
-ISSN={1662-5196},
-ABSTRACT={Analyses of brain function and anatomy using shared neuroimaging data is an important development, and have acquired the potential to be scaled up with the specification of a new Brain Imaging Data Structure (BIDS) standard. To date, a variety of software tools help researchers in converting their source data to BIDS but often require programming skills or are tailored to specific institutes, data sets, or data formats. In this paper, we introduce BIDScoin, a cross-platform, flexible, and user-friendly converter that provides a graphical user interface (GUI) to help users finding their way in BIDS standard. BIDScoin does not require programming skills to be set up and used and supports plugins to extend their functionality. In this paper, we show its design and demonstrate how it can be applied to a downloadable tutorial data set. BIDScoin is distributed as free and open-source software to foster the community-driven effort to promote and facilitate the use of BIDS standard.}
-}
-
-@article{Li2016,
-  doi = {10.1016/j.jneumeth.2016.03.001},
-  url = {https://doi.org/10.1016%2Fj.jneumeth.2016.03.001},
-  year = 2016,
-  month = {may},
-  publisher = {Elsevier {BV}},
-  volume = {264},
-  pages = {47--56},
-  author = {Xiangrui Li and Paul S. Morgan and John Ashburner and Jolinda Smith and Christopher Rorden},
-  title = {The first step for neuroimaging data analysis: {DICOM} to {NIfTI} conversion},
-  journal = {Journal of Neuroscience Methods}
-}
-
-@software{BoreDcm2Bids2023,
-  author = {Boré, Arnaud and Guay, Samuel and Bedetti, Christophe and Meisler, Steven and GuenTher, Nick},
-  doi = {10.5281/zenodo.8342572},
-  month = aug,
-  title = {{Dcm2Bids}},
-  url = {https://github.com/UNFmontreal/Dcm2Bids},
-  version = {3.1.0},
-  year = {2023}
-}
-
[…]
