### 2023-11-22 #5575 @nbeliy
Dear @emdupre,

I made a release [1.5.0](https://github.com/CyclotronResearchCentre/bidsme/releases/tag/1.5.0), and uploaded on zendoo. Waiting the co-authors to check.

For the change of reference, I would need your help: changing to `[BIDS; @Gorgolewski2016]` made it appear in text as `[BIDS; Gorgolewski(2016)]`, see [here](https://github.com/CyclotronResearchCentre/bidsme/actions/runs/6955665422).
