### 2023-11-23 #5575 @nbeliy
The tagged version: [1.5.0](https://github.com/CyclotronResearchCentre/bidsme/releases/tag/1.5.0)

The archived version on [zenodo](https://zenodo.org/records/10185300):

 - [10.5281/zenodo.10185300](https://doi.org/10.5281/zenodo.10185300) -- for 1.5.0;
 - [10.5281/zenodo.7524615](https://zenodo.org/doi/10.5281/zenodo.7524615) -- for all versions.
