### 2023-11-29 #5575 @emdupre
Thank you, @nbeliy !

On the formatting front, [I think I've found the issue](https://stackoverflow.com/questions/75652085/problematic-semi-colon-in-prefix-for-in-text-citation). Could you please try escaping the semicolon ? So, either:

> [BIDS\; @Gorgolewski2016]

or, if that doesn't work:

> [BIDS$;$ @Gorgolewski2016]
