### 2023-11-29 #5575 @emdupre
On the archive, it looks like one author might have their first and last names flipped. Specifically, in the paper they appear as :

> Grégory Hammad

but in the archive they appear as :

>  Grégory, Hammad

implying that their full name is Hammad Grégory. Could you please confirm this ordering and align the other authors appropriately ?
