### 2023-11-29 #5575 @emdupre
Also on the archive, it looks like the license listed there is the GNU General Public License v3.0 or later. However, the license listed on the software is [the GNU General Public License v2.0](https://github.com/CyclotronResearchCentre/bidsme/blob/dev/LICENSE.md).

Could you please harmonize these ? You should be able to change the license on the archive by [editing its metadata](https://help.zenodo.org/docs/deposit/manage-records/#edit). Please let me know if this requires minting a new DOI.
