### 2023-12-02 #5575 @Kevin-Mattheus-Moerman
@nbeliy as the AEiC I will now help process the final steps. I have checked the paper, your repository, the archive link, and this review. All seems in order. I only have the below point that needs your attention:

- [ ] Check `developped`, `developpers`, `developper`, each should be written with a single `p`.
