### 2023-12-04 52f5692
Fixed typos in acknowledgements
